+++ .github/workflows/pr-ci.yml
@@ -55,6 +55,7 @@ jobs:
           # export RUSTC_WRAPPER=~/.cargo/bin/sccache
           # make test
           # make debug
+          export ENGINE_LABEL_VALUE=tiflash
           export RUST_BACKTRACE=full
           cargo check
           cargo test --package tests --test failpoints cases::test_normal
@@ -71,4 +72,5 @@ jobs:
           cargo test --package tests --test failpoints cases::test_merge
           cargo test --package tests --test failpoints cases::test_import_service
           cargo test --package tests --test failpoints cases::test_proxy_replica_read
-          cargo test --package tests --test failpoints cases::test_proxy
+          cargo test --features compat_new_proxy --package tests --test proxy test_normal
+          cargo test --package tests --test proxy test_proxy
