+++ Cargo.toml
@@ -58,6 +58,7 @@ cloud-storage-grpc = ["sst_importer/cloud-storage-grpc"]
 cloud-storage-dylib = ["sst_importer/cloud-storage-dylib"]
 pprof-dwarf = ["pprof/backtrace-rs"]
 pprof-fp = ["pprof/frame-pointer"]
+compat_new_proxy = ["engine_tiflash/compat_new_proxy"]
 
 # for testing configure propegate to other crates
 # https://stackoverflow.com/questions/41700543/can-we-share-test-utilites-between-crates
@@ -88,6 +89,7 @@ encryption_export = { path = "components/encryption/export", default-features = 
 engine_panic = { path = "components/engine_panic", default-features = false }
 engine_rocks = { path = "components/engine_rocks", default-features = false }
 engine_test = { path = "components/engine_test", default-features = false }
+engine_tiflash = { path = "engine_tiflash", default-features = false }
 engine_traits = { path = "components/engine_traits", default-features = false }
 engine_traits_tests = { path = "components/engine_traits_tests", default-features = false }
 error_code = { path = "components/error_code", default-features = false }
