+++ engine_tiflash/src/write_batch.rs
@@ -101,13 +101,17 @@ impl engine_traits::WriteBatch for RocksWriteBatch {
 }
 
 pub fn do_write(cf: &str, key: &[u8]) -> bool {
-    match cf {
-        engine_traits::CF_RAFT => true,
-        engine_traits::CF_DEFAULT => {
-            key == keys::PREPARE_BOOTSTRAP_KEY || key == keys::STORE_IDENT_KEY
-        }
-        _ => false,
+    #[cfg(feature = "compat_new_proxy")]
+    {
+        return match cf {
+            engine_traits::CF_RAFT => true,
+            engine_traits::CF_DEFAULT => {
+                key == keys::PREPARE_BOOTSTRAP_KEY || key == keys::STORE_IDENT_KEY
+            }
+            _ => false,
+        };
     }
+    return true;
 }
 
 impl RocksWriteBatch {
