+++ new-mock-engine-store/Cargo.toml
@@ -14,6 +14,7 @@ protobuf-codec = [
   "protobuf/bytes",
   "kvproto/protobuf-codec",
 ]
+compat_new_proxy = ["engine_tiflash/compat_new_proxy"]
 
 [dependencies]
 api_version = { path = "../components/api_version", default-features = false }
