+++ tests/Cargo.toml
@@ -48,6 +48,7 @@ snmalloc = ["tikv/snmalloc"]
 mem-profiling = ["tikv/mem-profiling"]
 sse = ["tikv/sse"]
 portable = ["tikv/portable"]
+compat_new_proxy = ["engine_tiflash/compat_new_proxy"]
 
 [dependencies]
 api_version = { path = "../components/api_version", default-features = false }
@@ -58,6 +59,7 @@ crc64fast = "0.1"
 crossbeam = "0.8"
 encryption = { path = "../components/encryption", default-features = false }
 engine_rocks_helper = { path = "../components/engine_rocks_helper" }
+engine_tiflash = { path = "../engine_tiflash", default-features = false }
 error_code = { path = "../components/error_code", default-features = false }
 fail = "0.5"
 file_system = { path = "../components/file_system" }
