+++ tests/failpoints/cases/mod.rs
@@ -16,7 +16,6 @@ mod test_metrics_overflow;
 mod test_normal;
 mod test_pd_client;
 mod test_pending_peers;
-mod test_proxy;
 mod test_rawkv;
 mod test_replica_read;
 mod test_replica_stale_read;
